desloppify: complete toast Ui/App ownership split
ToastConverters moved earlier (d3a4a74). Move the remaining VMs
(ToastItemViewModel, ToastActionViewModel) into CodeScope.Ui/ViewModels/
so the layer boundary lines up with the rest of the project: Ui owns
VMs and converters, App owns the host wiring (Popup HWND, View
templates, ToastService impl). XAML xmlns refs point at the Ui
assembly via assembly-qualified clr-namespace.
